Refactor API to use one integer type: int64_t
This should make the API simpler, and int64_t is enough to represent any integer
value we might need.

Range checks should be done inside the API functions, that way we can modify the
types of the actual fields/variables modified by the API without changes to the
API prototypes.

+++ src/nvim/api/defs.h
@@ -15,9 +15,9 @@ typedef struct {
   size_t size;
 } String;
 
-typedef uint16_t Buffer;
-typedef uint16_t Window;
-typedef uint16_t Tabpage;
+typedef int64_t Buffer;
+typedef int64_t Window;
+typedef int64_t Tabpage;
 
 typedef struct object Object;
 
@@ -27,7 +27,7 @@ typedef struct {
 } StringArray;
 
 typedef struct {
-  uint16_t row, col;
+  int64_t row, col;
 } Position;
 
 typedef struct {
